refactor(sdk): change PartitionId argument type to uint32 (#2009)
change PartitionId argument type to uint32

+++ foreign/go/contracts/partitions.go
@@ -67,9 +67,9 @@ func None() Partitioning {
 	}
 }
 
-func PartitionId(value int) Partitioning {
+func PartitionId(value uint32) Partitioning {
 	bytes := make([]byte, 4)
-	binary.LittleEndian.PutUint32(bytes, uint32(value))
+	binary.LittleEndian.PutUint32(bytes, value)
 
 	return Partitioning{
 		Kind:   PartitionIdKind,
